+++ composenav/src/main/java/com/adamkobus/compose/navigation/di/NavModuleInternalBinds.kt
@@ -1,8 +1,8 @@
 package com.adamkobus.compose.navigation.di
 
-import com.adamkobus.compose.navigation.NavActionConsumer
 import com.adamkobus.compose.navigation.NavActionVerifier
-import com.adamkobus.compose.navigation.model.NavActionConsumerImpl
+import com.adamkobus.compose.navigation.NavigationConsumer
+import com.adamkobus.compose.navigation.model.NavigationConsumerImpl
 import com.adamkobus.compose.navigation.model.StubActionVerifier
 import dagger.Binds
 import dagger.Module
@@ -15,7 +15,7 @@ import dagger.multibindings.IntoSet
 internal interface NavModuleInternalBinds {
 
     @Binds
-    fun bindsNavActionConsumer(impl: NavActionConsumerImpl): NavActionConsumer
+    fun bindsNavActionConsumer(impl: NavigationConsumerImpl): NavigationConsumer
 
     @Binds
     @IntoSet