+++ composenav/src/main/java/com/adamkobus/compose/navigation/error/NavIntentMappingTooDeepError.kt
@@ -1,0 +1,3 @@
+package com.adamkobus.compose.navigation.error
+
+class NavIntentMappingTooDeepError(message: String) : RuntimeException(message)