+++ composenav/src/main/java/com/adamkobus/compose/navigation/model/NavigationProcessor.kt
@@ -6,6 +6,8 @@ import com.adamkobus.compose.navigation.action.NavAction
 import com.adamkobus.compose.navigation.data.GlobalGraph
 import com.adamkobus.compose.navigation.data.NavGraph
 import com.adamkobus.compose.navigation.destination.CurrentDestination
+import com.adamkobus.compose.navigation.intent.NavIntent
+import com.adamkobus.compose.navigation.intent.NavIntentResolvingManager
 import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -22,7 +24,7 @@ internal class NavigationProcessor @Inject constructor(
     private val actionDispatcher: PendingActionDispatcher
 ) {
     private val scope: CoroutineScope = CoroutineScope(Dispatchers.IO + SupervisorJob())
-    private val actionBuffer = MutableSharedFlow<NavAction>()
+    private val actionBuffer = MutableSharedFlow<ProcessorTask>()
 
     private var onActionPerformed: CompletableDeferred<Unit>? = null
 
@@ -32,13 +34,16 @@ internal class NavigationProcessor @Inject constructor(
     private val currentDestination: CurrentDestination
         get() = destinationManager.currentDestination
 
+    private val navIntentResolver: NavIntentResolvingManager
+        get() = ComposeNavigation.getNavIntentResolvingManager()
+
     private val logger: NavLogger
         get() = ComposeNavigation.getLogger()
 
     init {
         scope.launch {
             actionBuffer.collect {
-                processAction(it)
+                processTask(it)
             }
         }
     }
@@ -49,7 +54,29 @@ internal class NavigationProcessor @Inject constructor(
 
     fun postNavAction(navAction: NavAction) {
         scope.launch {
-            actionBuffer.emit(navAction)
+            actionBuffer.emit(ProcessorTask.Action(navAction))
+        }
+    }
+
+    fun postNavIntent(intent: NavIntent) {
+        scope.launch {
+            actionBuffer.emit(ProcessorTask.Intent(intent))
+        }
+    }
+
+    private suspend fun processTask(task: ProcessorTask) {
+        when (task) {
+            is ProcessorTask.Action -> processAction(task.navAction)
+            is ProcessorTask.Intent -> processIntent(task.navIntent)
+        }
+    }
+
+    private suspend fun processIntent(navIntent: NavIntent) {
+        val action = navIntentResolver.resolve(navIntent, currentDestination)
+        if (action == null) {
+            logger.w("Intent $navIntent discarded, because it was not mapped to any action")
+        } else {
+            processAction(action)
         }
     }
 
@@ -80,5 +107,28 @@ internal class NavigationProcessor @Inject constructor(
         destinationManager.onBackStackUpdated(entry, backQueue)
         onActionPerformed?.complete(Unit)
         onActionPerformed = null
+    }
+}
+
+private sealed class ProcessorTask {
+
+    class Action(val navAction: NavAction) : ProcessorTask() {
+        override fun equals(other: Any?): Boolean {
+            return other is Action && other.navAction == navAction
+        }
+
+        override fun hashCode(): Int {
+            return navAction.hashCode()
+        }
+    }
+
+    class Intent(val navIntent: NavIntent) : ProcessorTask() {
+        override fun equals(other: Any?): Boolean {
+            return other is Intent && other.navIntent == navIntent
+        }
+
+        override fun hashCode(): Int {
+            return navIntent.hashCode()
+        }
     }
 }