+++ demo-common-ui/src/main/java/com/adamkobus/compose/navigation/demo/ui/nav/CatsBrowserGraph.kt
@@ -10,7 +10,7 @@ object CatsBrowserGraph : NavGraph("cats") {
 
     override fun startDestination(): NavDestination = CatsList
 
-    val CatsList = AppGraph.navDestination("catsList")
+    val CatsList = navDestination("catsList")
     val CatDetails = CatsList.next {
         param(PARAM_CAT_ID)
     }