+++ demo-common-ui/src/main/java/com/adamkobus/compose/navigation/demo/ui/nav/DogsBrowserGraph.kt
@@ -10,7 +10,7 @@ object DogsBrowserGraph : NavGraph("dogs") {
 
     override fun startDestination(): NavDestination = DogsList
 
-    val DogsList = AppGraph.navDestination("dogsList")
+    val DogsList = navDestination("dogsList")
     val DogDetails = DogsList.next {
         param(PARAM_DOG_ID)
     }