+++ demo-ui-dogs-browser/src/main/java/com/adamkobus/compose/navigation/demo/ui/demodialog/DemoDialogVM.kt
@@ -2,7 +2,7 @@ package com.adamkobus.compose.navigation.demo.ui.demodialog
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.adamkobus.compose.navigation.NavActionConsumer
+import com.adamkobus.compose.navigation.NavigationConsumer
 import com.adamkobus.compose.navigation.demo.ui.nav.FromDemoDialog
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
@@ -10,13 +10,13 @@ import javax.inject.Inject
 
 @HiltViewModel
 class DemoDialogVM @Inject constructor(
-    private val navActionConsumer: NavActionConsumer
+    private val navigationConsumer: NavigationConsumer
 ) : ViewModel() {
 
     val interactions = DemoDialogInteractions(
         onCancelClicked = {
             viewModelScope.launch {
-                navActionConsumer.offer(FromDemoDialog.Dismiss)
+                navigationConsumer.offer(FromDemoDialog.Dismiss)
             }
         }
     )